Added a configuration option to suffix/prefix the prompt
Now, you can add a prefix and/or suffix to your git-radar prompt that
will only show if you are in a git repo.

Note: The PROMPT_FORMAT line was changed to remove the leading space in
the string. With it in, if you define a prefix a space will still be
there even if the user may not want it. So it can be overridden by the
user with one of the new variables, but by default, the prefix still
adds that space anyways.